+++ src/question-one/QuestionOne.js
@@ -3,7 +3,7 @@ import React, {useState, useEffect} from 'react';
 import { SectionGroup } from '../components/section/SectionGroup'
 import { SectionPanel } from '../components/section/SectionPanel'
 
-import SearchableSelectField from './SearchableSelectField'
+import SearchableSelectField from '../components/common/SearchableSelectField'
 import colors from '../style/colors'
 
 import './QuestionOne.css'
